+++ R/api.R
@@ -28,6 +28,9 @@ api_build <- function(queue, validate = FALSE, logger = NULL) {
   api$handle(endpoint_rehydrate_submit(queue))
   api$handle(endpoint_rehydrate_status(queue))
   api$handle(endpoint_rehydrate_result(queue))
+  api$handle(endpoint_upload_input(queue))
+  api$handle(endpoint_upload_output(queue))
+  api$handle(endpoint_prerun(queue))
   api$handle(endpoint_hintr_version())
   api$handle(endpoint_hintr_worker_status(queue))
   api$handle(endpoint_hintr_stop(queue))
@@ -53,16 +56,17 @@ api_postserialize <- function(data, req, res, value) {
 #' if NULL
 #' @param workers Number of workers to spawn
 #' @param results_dir The dir for results to be saved to
-#' @param prerun_dir The directory to store prerun results
+#' @param inputs_dir THe directory where input files are stored
 #' @param log_level The "lgr" log level to use
 #'
 #' @return Running API
 #' @export
 api <- function(queue_id = NULL, workers = 2,
-                results_dir = tempdir(), prerun_dir = NULL,
+                results_dir = tempdir(), inputs_dir = NULL,
                 log_level = "info") {
-  queue <- Queue$new(queue_id, workers, results_dir = results_dir,
-                     prerun_dir = prerun_dir)
+  queue <- Queue$new(queue_id, workers,
+                     results_dir = results_dir,
+                     inputs_dir = inputs_dir)
   queue$queue$worker_delete_exited()
   logger <- porcelain::porcelain_logger(log_level)
   api_build(queue, logger = logger)
@@ -363,6 +367,41 @@ endpoint_adr_metadata <- function(queue) {
   porcelain::porcelain_endpoint$new("GET",
                                     "/meta/adr/<id>",
                                     adr_metadata(queue),
+                                    returning = response)
+}
+
+endpoint_upload_input <- function(queue) {
+  input <- porcelain::porcelain_input_body_binary("file")
+  response <- porcelain::porcelain_returning_json(
+    "File.schema", schema_root())
+  porcelain::porcelain_endpoint$new("POST",
+                                    "/internal/upload/input/<filename>",
+                                    upload_file(queue$inputs_dir),
+                                    input,
+                                    returning = response)
+}
+
+endpoint_upload_output <- function(queue) {
+  input <- porcelain::porcelain_input_body_binary("file")
+  response <- porcelain::porcelain_returning_json(
+    "File.schema", schema_root())
+  porcelain::porcelain_endpoint$new("POST",
+                                    "/internal/upload/result/<filename>",
+                                    upload_file(queue$results_dir),
+                                    input,
+                                    returning = response)
+}
+
+endpoint_prerun <- function(queue) {
+  input <- porcelain::porcelain_input_body_json("input",
+                                                "PrerunRequest.schema",
+                                                schema_root())
+  response <- porcelain::porcelain_returning_json(
+    "ProjectState.schema.json", schema_root())
+  porcelain::porcelain_endpoint$new("POST",
+                                    "/internal/prerun",
+                                    prerun(queue),
+                                    input,
                                     returning = response)
 }