+++ tests/testthat/test-main.R
@@ -6,12 +6,11 @@ test_that("main_api_args", {
   expect_null(default$queue_id)
   expect_equal(default$workers, 2)
   expect_true(file.exists(default$results_dir))
-  expect_true(file.exists(default$prerun_dir))
   expect_equal(
     main_api_args(c("--workers=0", "--port", "80", "--results-dir=out",
-                    "--prerun-dir=pr", "hintr")),
+                    "--prerun-dir=pr", "--inputs-dir=in", "hintr")),
     list(port = 80, queue_id = "hintr", workers = 0, results_dir = "out",
-         prerun_dir = "pr"))
+         inputs_dir = "in"))
 })
 
 test_that("main_worker_args", {